pyproject.toml
[tool.poetry]
name = "dyndis"
version = "0.2.0"
description = ""
authors = ["ben avrahami <dontcallme@illcall.you>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/bentheiii/dyndis"
packages = [
    {include="dyndis"}
]

[tool.poetry.dependencies]
python = "^3.7"
sortedcontainers = "*"

[tool.poetry.dev-dependencies]
pydantic = "*"
flake8 = { version = "*", allow-prereleases = true }
pytest-cov = "*"
#sphinx = "*"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

